requirements.txt
flask
certifi
requests
messagebird
pytz
docxtpl
uwsgi
py-dateutil
netifaces
mechanize
netifaces